Fixes for 64-bit, compiler warnings, code modernisation and new build system

+++ util/transname.h
@@ -1,1 +1,1 @@
-void transname(char *name, char *namebuf, int n);
+void transname(char *name, char *namebuf, size_t n);
